python_service/requirements.txt
# Python ML service dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
pytest>=8.0.0
httpx>=0.27.0
